Split CWS' handlers
